#2696 - Document-level recommendations
- Pull code to reject and skip suggestions up into the base class so that metadata and relation suggestion supports automatically get them
